Working with segments

+++ source/ResourceManager/zlib/uncompr.c
@@ -37,8 +37,7 @@ int ZEXPORT uncompress (dest, destLen, source, sourceLen)
 
     stream.next_out = dest;
     stream.avail_out = (uInt)*destLen;
-    
-    if ((uLong)stream.avail_out != (int)*destLen) return Z_BUF_ERROR;
+    if ((uLong)stream.avail_out != *destLen) return Z_BUF_ERROR;
 
     stream.zalloc = (alloc_func)0;
     stream.zfree = (free_func)0;
